Move report source presentation helper
