+++ apps/server/vibesensor/use_cases/history/report_document/_candidate_resolver.py
@@ -7,10 +7,10 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from vibesensor.domain import Finding, TestRun
-from vibesensor.report_i18n import human_source
 from vibesensor.shared.boundaries.reporting.confidence_facts import ReportConfidenceFacts
 from vibesensor.shared.report_presentation import (
     confidence_reason_text,
+    human_source,
     strength_label,
     strength_text,
 )
